Make Named trait a part of generic trees
